Fix duplication of sections across cards
Currently each event gets added as a section to a card so old alerts
are coming through again and again on new cards.

This bug can be replicated by taking the example test json and running
it back to back a few times. The first card will have 1 section, the
second two sections (1 duplicate), the third 3 sections (2 duplicates),
and so on.