support authorization header for dynamicwebhook
This is intended to work with Alertmanagers credential handling, where
it automatically "hides" pieces of information that are considered
secret. As an MS-Teams Incoming Webhook acts as an authorization token
it can be considered as such a credential.
Alertmanager however does not consider webhook URLs to be credential
like information. If we store the webhook URL in the authorization
header instead it is treated like a credential by Alertmanager.